+++ stem/util/proc.py
@@ -333,8 +333,8 @@ def connections(pid):
   :param int pid: process id of the process to be queried
 
   :returns: A listing of connection tuples of the form **[(local_ipAddr1,
-    local_port1, foreign_ipAddr1, foreign_port1, protocol), ...]** (addresses
-    and protocols are strings and ports are ints)
+    local_port1, foreign_ipAddr1, foreign_port1, protocol, is_ipv6), ...]**
+    (addresses and protocols are strings and ports are ints)
 
   :raises: **IOError** if it can't be determined
   """
@@ -403,7 +403,7 @@ def connections(pid):
           local_ip, local_port = _decode_proc_address_encoding(l_addr)
           foreign_ip, foreign_port = _decode_proc_address_encoding(f_addr)
           protocol = proc_file_path[10:]
-          conn.append((local_ip, local_port, foreign_ip, foreign_port, protocol))
+          conn.append((local_ip, local_port, foreign_ip, foreign_port, protocol, False))
 
       proc_file.close()
     except IOError as exc: